Editor Settings Working WIth Behavior Editor
Finally created a workaround for the issue by creating a different set of variables and then using a method to assign them in SIsEditor.cs rather than using the
'new' keyword to supress the original variables.

Editor could use some polishing on using the Unity asset finder as it cannot search for parents that are generic, but I'm moving on and putting that issue on the backburner.